build: include `ext/manifest.json` as-is if it exists
The build script generates `manifest.json` which contains the version
number, but it fails when building from a tarball because the archive
contains no Git tag information.
I've had some success by using a `.yarn.env` file to emulate the GitHub CI
environment, which was a bit hacky, though. It's much simpler to let the
user provide `manifest.json` themself.
If `ext/manifest.json` is present, it will now be copied as-is; otherwise
a dummy version name with a timestamp will be generated, and a warning
will be shown.